Move Provider DTOs to a separate folder
