+++ OutOfSchool/OutOfSchool.WebApi/Controllers/V1/ProviderController.cs
@@ -14,7 +14,7 @@ using OutOfSchool.Common;
 using OutOfSchool.Common.Extensions;
 using OutOfSchool.Common.PermissionsModule;
 using OutOfSchool.WebApi.Common;
-using OutOfSchool.WebApi.Models;
+using OutOfSchool.WebApi.Models.Providers;
 using OutOfSchool.WebApi.Services;
 
 namespace OutOfSchool.WebApi.Controllers.V1
