+++ OutOfSchool/OutOfSchool.WebApi/Models/Providers/ProviderDto.cs
@@ -10,7 +10,7 @@ using OutOfSchool.Services.Enums;
 using OutOfSchool.WebApi.Models.SubordinationStructure;
 using OutOfSchool.WebApi.Util.JsonTools;
 
-namespace OutOfSchool.WebApi.Models
+namespace OutOfSchool.WebApi.Models.Providers
 {
     public class ProviderDto
     {
@@ -90,7 +90,7 @@ namespace OutOfSchool.WebApi.Models
         [MaxLength(256)]
         public string CoverImageId { get; set; } = string.Empty;
 
-        [JsonProperty(NullValueHandling=NullValueHandling.Ignore)]
+        [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
         public IFormFile CoverImage { get; set; }
 
         [ModelBinder(BinderType = typeof(JsonModelBinder))]
