+++ OutOfSchool/OutOfSchool.WebApi/Models/Providers/ProviderLicenseStatusDto.cs
@@ -2,14 +2,14 @@
 using System.ComponentModel.DataAnnotations;
 using OutOfSchool.Services.Enums;
 
-namespace OutOfSchool.WebApi.Models
+namespace OutOfSchool.WebApi.Models.Providers
 {
-    public class ProviderStatusDto
+    public class ProviderLicenseStatusDto
     {
         [Required]
         public Guid ProviderId { get; set; }
 
         [Required]
-        public ProviderStatus Status { get; set; }
+        public ProviderLicenseStatus LicenseStatus { get; set; }
     }
 }
