+++ OutOfSchool/OutOfSchool.WebApi/Models/Providers/ProviderStatusDto.cs
@@ -2,14 +2,14 @@
 using System.ComponentModel.DataAnnotations;
 using OutOfSchool.Services.Enums;
 
-namespace OutOfSchool.WebApi.Models
+namespace OutOfSchool.WebApi.Models.Providers
 {
-    public class ProviderLicenseStatusDto
+    public class ProviderStatusDto
     {
         [Required]
         public Guid ProviderId { get; set; }
 
         [Required]
-        public ProviderLicenseStatus LicenseStatus { get; set; }
+        public ProviderStatus Status { get; set; }
     }
 }
