+++ OutOfSchool/OutOfSchool.WebApi/Services/IProviderServiceV2.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Threading.Tasks;
-using OutOfSchool.WebApi.Models;
+using OutOfSchool.WebApi.Models.Providers;
 
 namespace OutOfSchool.WebApi.Services
 {
