+++ OutOfSchool/Tests/OutOfSchool.Tests.Common/TestDataGenerators/ProviderDtoGenerator.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using Bogus;
 using OutOfSchool.Common.Enums;
 using OutOfSchool.Services.Enums;
-using OutOfSchool.WebApi.Models;
+using OutOfSchool.WebApi.Models.Providers;
 
 namespace OutOfSchool.Tests.Common.TestDataGenerators
 {
